added readme.md and renamed folders
